+++ sensing/autoware_downsample_filters/src/voxel_grid_downsample_filter/voxel_grid_downsample_filter_node.cpp
@@ -97,7 +97,9 @@ void VoxelGridDownsampleFilter::input_callback(const PointCloud2ConstPtr cloud)
 
 bool VoxelGridDownsampleFilter::is_valid(const PointCloud2ConstPtr & cloud)
 {
-  if (cloud->width * cloud->height * cloud->point_step != cloud->data.size()) {
+  if (
+    static_cast<std::size_t>(cloud->width) * cloud->height * cloud->point_step !=
+    cloud->data.size()) {
     RCLCPP_WARN(
       this->get_logger(),
       "Invalid PointCloud (data = %zu, width = %d, height = %d, step = %d) with stamp %f, "
